requirements.txt
mcp
openai
jikanpy-v4
fastapi
watchfiles
uvicorn[standard]
opencv-python
pyserial